Added Description/ Approch
The person toggles(on/off) the switch every time he goes to check the corridor.
Every time he toggles only that switch which divides its i iteration.
Means - 
For example - 4 bulbs 
i(iteration)    1              2         3          4           factors(i)
          On         off       off        off            1
          Off         on       off        off           1, 2 
          On         on       on        off            1,3
          Off         off       on        on            1, 2, 4

Therefore it is observed that to determine if the bulb is on/off we have to count the number of toggles of that switch. 

To count the toggle of the switch we need to know the total number of factors of the i tern.

If the number of factors of the ith term is an odd number then the bulb will be on else the bulb will be off.

To check the total number of factors are odd or even there is test which is 
    If number of factors of the number are odd than the int(sqrt(n) * int(sqrt(n)) = n(perfect square) else the number of factors are even.

 Now in an over situation if the number of factors are odd then the bulb is ON else the bulb is off.

Therefore 
Int x = sqrt(n)
    if the x*x==n:
   than the bulb is “on”
           else :
              bulb is “off” 

1. Light more Light -

Our task - to check if the last bulb(nth) is on or off. 

How to check if the last bulb is on or off.

int x = sqrt(n)
if the x*x==n:
   than the bulb is “on”
           else :
              bulb is “off”