Added Description/ Approach
2. Carmichael Number - 


Fermat test is the test to check if a big number is a prime number or no  prime number.
Fermat test equation -  (a^n) % n = a ,     
a = (1, n-1) 
if the number n passes the test several times then there are high chances that the number is prime. 
Drawback - some numbers which are not prime can also pass this test such numbers are known as carmichael numbers.


Our task - to print yes/no depending upon if the entered number is carmichael number or not.

How to check if the number is carmichael number or not -

Check if the number is prime or not. 
 If the number is prime then return false that means it is a normal number and not a carmichael number, else if it is not prime then move to step 2.  
If the number is not prime then check for the fermat test equation. 
 If the number do not satisfy the equation then the number is not carmichael number,
 else it is a carmichael number.