Added Description/Approach
It is based on a backtracking algorithm. 

Here, we repetitively do call the function gcd with value(b, b%a)
Where b is always a greater number than a. Otherwise we will swap the values of a and b.

After finding the final value of gcd we calculate the value of x & y such that gcd= Ax + By.
And finally we return the value of x, y and z.
While finding the value of x & y x = x - (a/b)*y