Added Description/Approach
Example: you have a number 11 and 4 prime number is 2, 3, 3, 3. Another Example you have
10 so this time your 4 prime numbers are 2, 2, 3, 3.

so this time fist we can check, our first number is odd or even.
if we show that our number is even then we can print 2 prime numbers first prime number is 2
and the second prime number is 2. and then we decrease 4 from our number,

if we show that our number is odd then we can print 2 prime numbers first prime number is 2
and the second prime number is 3. and then we decrease 5 from our number,

Now we need to just find 2 prime number so now our work is easy. First of all, we can find 1 to
10000000 all prime numbers (Using Sieve ) and save all prime numbers on an array or vector.

So, now we take a prime number (3rd prime) from our array or vector and find our last / 4th
number is our number - 3rd prime number.

Then we need to check 4th number is prime or not. if our 4th number is prime this you need to
check 3rd and 4th number sum is input number. otherwise, you need to take another prime
number from the array or vector and check the same way.

And finally, we check our 3rd number and 4th number sum is input number, this time we need to
print this 3rd number and 4th number and BREAK our loop.