Update Bicoloring.cpp
Approach:

In 1976 the “Four Color Map Theorem” was proven with the assistance of a computer. This theorem states that every map can be colored using only four colors, in such a way that no region is colored using the same color as a neighbor region.

Here you are asked to solve a simpler similar problem. You have to decide whether a given arbitrary connected graph can be bicolored. That is, if one can assign colors (from a palette of two) to the nodes in such a way that no two adjacent nodes have the same color. To simplify the problem you can assume:
no node will have an edge to itself.
the graph is nondirected. That is, if a node a is said to be connected to a node b, then you must assume that b is connected toa.
the graph will be strongly connected. That is, there will be at least one path from any node to any other node.

Method One: BFS wide search:

It is seen from the subject, for each node, and all its contact point, and this point must not the same color. So, it is natural wide search can do: Choose one of the points, the points assigned to one color, you can use the number 0 instead, then wide search, and then all of his adjacent points can be assigned to another colors, can be replaced by 1. So, the search continues, if you encounter a point that has been assigned over, it would be judged, he has some value and this is not to give it the same value, if it is the same, to continue. If different, then the judge is not direct.

Method Two: DFS deep search

Again, this problem can also be used to do a deep search. The basic idea is deep search, the search continues to go in one direction, did not walk a step of dyeing, this is the opposite of the current color and a bit of color. If the search to a dyed (that is loop), it can also be judged, it is not and has some color to it the color is consistent. Inconsistent, then it cannot be judged.