Create Freckles.java
I used Prim's algorithm to solve this problem.

First, we need to calculate the distance between the freckles. Then, call the Prim method, which will link the freckles, storing each edge from the current freckle in a PriorityQueue. This queue sorts the edges in order to use those that have the smallest distances first. Once all the freckles are linked, the method returns the sum of the distance between the freckles.